pyproject.toml
[project]
name = "memory-ai"
version = "0.1.0"
description = "Personal persistent memory system for LLMs"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "litellm>=1.40",
    "instructor>=1.0",
    "pydantic>=2.0",
    "mcp[cli]>=1.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
memory = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
